refactor: auth views into partials
- Allows partials to be used on auth pages and modal
- Allows for swapping views or route change
