+++ src/components/app/modal/components/body/body.module.scss
@@ -1,3 +1,4 @@
 .body {
-  background-color: transparent;
+  flex: 1;
+  overflow: auto;
 }
