+++ src/components/app/modal/components/footer/footer.tsx
@@ -1,15 +1,13 @@
 import React from 'react';
+import cx from 'classnames';
 import { ModalContext } from '../../modal-context';
+import styles from './footer.module.scss';
 
-const Footer = ({ children }: any) => {
+const Footer = ({ className, children }: any) => {
   return (
     <ModalContext.Consumer>
       {() => {
-        return (
-          <div>
-            <div>{children}</div>
-          </div>
-        );
+        return <div className={cx(styles.footer, className)}>{children}</div>;
       }}
     </ModalContext.Consumer>
   );
