+++ src/components/app/modal/components/header/header.tsx
@@ -1,15 +1,13 @@
 import React from 'react';
+import cx from 'classnames';
 import { ModalContext } from '../../modal-context';
+import styles from './header.module.scss';
 
-const Header = ({ children }: any) => {
+const Header = ({ className, children }: any) => {
   return (
     <ModalContext.Consumer>
       {() => {
-        return (
-          <div>
-            <div>{children}</div>
-          </div>
-        );
+        return <div className={cx(styles.header, className)}>{children}</div>;
       }}
     </ModalContext.Consumer>
   );
